version 0.3.0 Snapshot

+++ Gestion_Ressources_materielles/src/main/java/org/pi/consumer/UserRepository.java
@@ -8,12 +8,12 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 	
 	public User findByEmail(String email);
     
     @Query("select role as role from User where email=?1")
     public List<String> findRoleByEmail(String email);
     
-    
+
 }
